Fixed type

+++ android/src/main/java/com/github/wuxudong/rncharts/charts/BarLineChartBaseManager.java
@@ -69,7 +69,7 @@ public abstract class BarLineChartBaseManager<T extends BarLineChartBase, U exte
     }
 
     @ReactProp(name = "maxHighlightDistance")
-    public void setMaxHighlightDistance(BarLineChartBase chart, Integer maxHighlightDistance) {
+    public void setMaxHighlightDistance(BarLineChartBase chart, float maxHighlightDistance) {
         chart.setMaxHighlightDistance(maxHighlightDistance);
     }
